+++ src/reanimated2/layoutReanimation/defaultAnimations/Stretch.ts
@@ -1,0 +1,97 @@
+import { BaseAnimationBuilder } from '../defaultAnimationsBuilder';
+
+export class StretchInX extends BaseAnimationBuilder {
+  static createInstance() {
+    return new StretchInX();
+  }
+
+  build() {
+    const delayFunction = this.getDelayFunction();
+    const [animation, config] = this.getAnimationAndConfig();
+    const delay = this.delayV;
+
+    return () => {
+      'worklet';
+      return {
+        animations: {
+          transform: [{ scaleX: delayFunction(delay, animation(1, config)) }],
+        },
+        initialValues: {
+          transform: [{ scaleX: 0 }],
+        },
+      };
+    };
+  }
+}
+
+export class StretchInY extends BaseAnimationBuilder {
+  static createInstance() {
+    return new StretchInY();
+  }
+
+  build() {
+    const delayFunction = this.getDelayFunction();
+    const [animation, config] = this.getAnimationAndConfig();
+    const delay = this.delayV;
+
+    return () => {
+      'worklet';
+      return {
+        animations: {
+          transform: [{ scaleY: delayFunction(delay, animation(1, config)) }],
+        },
+        initialValues: {
+          transform: [{ scaleY: 0 }],
+        },
+      };
+    };
+  }
+}
+
+export class StretchOutX extends BaseAnimationBuilder {
+  static createInstance() {
+    return new StretchOutX();
+  }
+
+  build() {
+    const delayFunction = this.getDelayFunction();
+    const [animation, config] = this.getAnimationAndConfig();
+    const delay = this.delayV;
+
+    return () => {
+      'worklet';
+      return {
+        animations: {
+          transform: [{ scaleX: delayFunction(delay, animation(0, config)) }],
+        },
+        initialValues: {
+          transform: [{ scaleX: 1 }],
+        },
+      };
+    };
+  }
+}
+
+export class StretchOutY extends BaseAnimationBuilder {
+  static createInstance() {
+    return new StretchOutY();
+  }
+
+  build() {
+    const delayFunction = this.getDelayFunction();
+    const [animation, config] = this.getAnimationAndConfig();
+    const delay = this.delayV;
+
+    return () => {
+      'worklet';
+      return {
+        animations: {
+          transform: [{ scaleY: delayFunction(delay, animation(0, config)) }],
+        },
+        initialValues: {
+          transform: [{ scaleY: 1 }],
+        },
+      };
+    };
+  }
+}
