+++ src/reanimated2/layoutReanimation/defaultAnimations/index.js
@@ -1,0 +1,5 @@
+export * from './Flip';
+export * from './Stretch';
+export * from './Fade';
+export * from './Slide';
+export * from './Zoom';
